Fix tree() method Add tests for tree Move files fixture to conftest.py
