+++ tests/pipeline/test_resources.py
@@ -1,20 +1,5 @@
 from src.pipeline.resources import SFTPResource
 import pytest
-import os
-
-
-@pytest.fixture
-def files():
-    return {
-        'root_file.dat': 'root file contents',
-        'tmp': {},
-        'a_dir': {'somefile.txt': "File content"},
-        'listed_dir': {
-            'a': 'a_contents',
-            'b': 'b_contents',
-            'subdir': {'sub_c': 'sub_c_contents', 'subsubdir': {'subsub_d': 'subsub_d_contents'}},
-        },
-    }
 
 
 class TestSFTPResource:
@@ -22,7 +7,6 @@ class TestSFTPResource:
     def test_open_close_connection(sftpserver, files):
         with sftpserver.serve_content(files):
             res = SFTPResource(hostname=sftpserver.host, port=sftpserver.port, username='user', password='pw')
-            print(res)
             assert res.connected is False
             res.connect()
             assert res.connected is True
@@ -30,51 +14,27 @@ class TestSFTPResource:
             assert res.connected is False
 
     @staticmethod
-    @pytest.mark.parametrize(
-        'local_filepath, remote_filepath, mkdirs',
-        [
-            ('somefile.txt', '/tmp/somefile.txt', False),
-            ('otherfile.txt', '/totally/random/path/mattfile', True),
-            ('a_poem', '/1_file/2_file/red_file/blue_file/fish.py', True),
-            ('important_documents', '/listed_dir/subdir/subsubdir/stash_here', False),
-        ],
-    )
-    def test_put_file(sftpserver, files, local_filepath, remote_filepath, mkdirs, tmp_path):
+    def test_put_file(sftpserver, files):
         with sftpserver.serve_content(files):
             with SFTPResource(
                 hostname=sftpserver.host, port=sftpserver.port, username='user', password='pw'
             ) as sftp_resource:
-                local_filepath = os.path.join(tmp_path, local_filepath)
-                assert not sftp_resource.exists(remote_filepath)
-
-                with open(local_filepath, 'w+') as f:
+                assert not sftp_resource.exists('/tmp/somefile.txt')
+                with open('/tmp/somefile.txt', 'w+') as f:
                     f.write('This is a file!')
-
-                sftp_resource.put_file(from_path=local_filepath, to_path=remote_filepath, mkdirs=mkdirs)
-
-                assert sftp_resource.exists(remote_filepath)
+                sftp_resource.put_file('/tmp/somefile.txt')
+                assert sftp_resource.exists('/tmp/somefile.txt')
 
     @staticmethod
-    @pytest.mark.parametrize(
-        'remote_filepath, local_filepath, mkdirs',
-        [
-            ('/a_dir/somefile.txt', 'tmp/somefile.txt', True),
-            ('/listed_dir/subdir/sub_c', 'mattfile', False),
-            ('/listed_dir/a', '1_file/2_file/red_file/blue_file/fish.py', True),
-            ('/root_file.dat', 'listed_dir/subdir/subsubdir/stash_here', True),
-        ],
-    )
-    def test_get_file(sftpserver, files, remote_filepath, local_filepath, mkdirs, tmp_path):
+    def test_get_file(sftpserver, files):
         with sftpserver.serve_content(files):
             with SFTPResource(
                 hostname=sftpserver.host, port=sftpserver.port, username='user', password='pw'
             ) as sftp_resource:
-                local_filepath = os.path.join(tmp_path, local_filepath)
-                assert not os.path.isfile(local_filepath)
 
-                sftp_resource.get_file(remote_filepath, local_filepath)
-
-                assert os.path.isfile(local_filepath)
+                assert sftp_resource.exists('/a_dir/somefile.txt')
+                sftp_resource.put_file('/tmp/somefile.txt')
+                assert sftp_resource.exists('/tmp/somefile.txt')
 
     @staticmethod
     @pytest.mark.parametrize(
@@ -130,10 +90,17 @@ class TestSFTPResource:
                 assert [x.filename for x in sftp_resource.ls(**ls_params)] == expected_output
 
     @staticmethod
-    @pytest.mark.skip(reason='Unfinished - hangs near completion - fix tree()?')
-    def test_tree(sftpserver, files):
+    @pytest.mark.parametrize(
+        'tree_params, expected_output',
+        [
+            ({'path': 'listed_dir'}, ['a', 'b', 'sub_c', 'subsub_d', 'i', 'a', 'of']),
+            ({'path': 'listed_dir/subdir'}, ['sub_c', 'subsub_d']),
+            ({'path': 'listed_dir/other_subdir'}, ['i', 'a', 'of']),
+        ],
+    )
+    def test_tree(sftpserver, files, tree_params, expected_output):
         with sftpserver.serve_content(files):
             with SFTPResource(
                 hostname=sftpserver.host, port=sftpserver.port, username='user', password='pw'
             ) as sftp_resource:
-                sftp_resource.tree('listed_dir')
+                assert [x.filename for x in sftp_resource.tree(**tree_params)] == expected_output
